refine findOrCreateUserByAccount
